code clean and imagery
