+++ src/context/context.tsx
@@ -1,4 +1,4 @@
-import React, { createContext, useReducer, Dispatch, ReactNode } from "react";
+import React, { createContext, useReducer, Dispatch } from "react";
 import { 
   productReducer, 
   shoppingCartReducer, 
@@ -6,31 +6,16 @@ import {
   ProductActions, 
   ShoppingCartActions, 
   ShoppingCartProductsActions, 
-  ShoppingCartProducts 
 } from "./reducer";
 
-type ProductType = {
-  [x: string]: ReactNode;
-  id: number;
-  name: string;
-  price: number;
-  description?:string;
-  quantity?: number;
-  category?: string;
-};
+import { InitialStateType, ShoppingCartProducts } from "../types/types";
 
-
-type InitialStateType = {
-  products: ProductType[];
-  shoppingCart: number;
-  shoppingCartProducts: ShoppingCartProducts;  // products in shopping cart
-};
 
 const initialState = {
   products: [{
     "id": 201,
     "name": "FaceMaskOne",
-    "image": "/images/pic1.jpg",
+    "image": "pic1.jpg",
     "price": 5.0,
     "category": "basic",
     "quantity": 5
@@ -38,7 +23,7 @@ const initialState = {
   {
     "id": 202,
     "name": "FaceMaskOne",
-    "image": "/images/pic2.jpg",
+    "image": "/pic2.jpg",
     "price": 6.0,
     "category": "basic",
     "quantity": 5
@@ -46,7 +31,7 @@ const initialState = {
   {
     "id": 203,
     "name": "FaceMaskOne",
-    "image": "/images/pic3.jpg",
+    "image": "/pic3.jpg",
     "price": 11.0,
     "category": "full" ,
     "quantity": 5
@@ -54,7 +39,7 @@ const initialState = {
   {
     "id": 204,
      "name": "FaceMaskOne",
-     "image": "/images/pic4.jpg",
+     "image": "/pic4.jpg",
      "price": 120.0,
      "category": "full" ,
      "quantity": 5
@@ -62,7 +47,7 @@ const initialState = {
   {
     "id": 205,
     "name": "FaceMaskOne",
-    "image": "/images/pic5.jpg",
+    "image": "/pic5.jpg",
     "price": 8.0,
     "category": "half" ,
     "quantity": 5
