--- src/types/productType.ts
@@ -1,1 +1,0 @@
-export type Product = { name: string; price: number; id: number, description: string; image: string };
